Nixsecura's own ctf machine
This CTF machine was created by Imran Guhar Khatib, the founder of NixSecura Services. It is a custom-built, offline challenge designed specifically for learning and practicing cybersecurity skills. Since this machine is not available online, it offers a unique hands-on experience for developing real-world offensive and defensive techniques such as bruteforcing